🐛 removed mesakit_home environment variable references
